fix: ClonedPlugins need to support signed values for priority

+++ kong/cloned_plugin_definition.go
@@ -5,7 +5,7 @@ package kong
 type ClonedPluginDefinition struct {
 	ID        *string   `json:"id,omitempty" yaml:"id,omitempty"`
 	Name      *string   `json:"name,omitempty" yaml:"name,omitempty"`
-	Priority  *uint64   `json:"priority,omitempty" yaml:"priority,omitempty"`
+	Priority  *int      `json:"priority,omitempty" yaml:"priority,omitempty"`
 	Ref       *string   `json:"ref,omitempty" yaml:"ref,omitempty"`
 	CreatedAt *int64    `json:"created_at,omitempty" yaml:"created_at,omitempty"`
 	UpdatedAt *int64    `json:"updated_at,omitempty" yaml:"updated_at,omitempty"`
